Changed math.h functions to use single precision float versions

+++ src/ExponentialResponse.cpp
@@ -15,7 +15,7 @@ float ExponentialResponse::response (const float value, const float minValue, co
 {
 	const float span = maxValue - minValue;
 	const float normalizedValue = value / span;
-	const float normalized = ( pow(m_Slope + 1, normalizedValue) - 1.0f ) / m_Slope;
+	const float normalized = ( powf(m_Slope + 1, normalizedValue) - 1.0f ) / m_Slope;
 
 	return std::fmin( ((normalized * span) - std::abs(minValue)), maxValue );
 }
